requirements.txt
flask
paramiko
python-dotenv
Flask-Cors
pystray
pillow
zeroconf